# contract/openapi.yaml
openapi: 3.0.0
info:
  title: Kafbat UI Api Service
  contact:
    name: API Support
    email: support@kafbat.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.2.0
...
paths:
...
  /api/clusters/{clusterName}/topics/{topicName}/analysis:
    get:
      operationId: getTopicAnalysis
      summary: getTopicAnalysis
      parameters:
        - name: clusterName
          in: path
          required: true
          schema:
            type: string
        - name: topicName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicAnalysis'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiNotFoundResponse'
      tags:
        - Topics
    post:
      operationId: analyzeTopic
      summary: analyzeTopic
      parameters:
        - name: clusterName
          in: path
          required: true
          schema:
            type: string
        - name: topicName
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiNotFoundResponse'
      tags:
        - Topics
    delete:
      operationId: cancelTopicAnalysis
      summary: cancelTopicAnalysis
      parameters:
        - name: clusterName
          in: path
          required: true
          schema:
            type: string
        - name: topicName
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiNotFoundResponse'
      tags:
        - Topics
...
components:
  schemas:
...
    ApiNotFoundResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
      description: ''
...
    TopicAnalysis:
      type: object
      properties:
        progress:
          $ref: '#/components/schemas/TopicAnalysisProgress'
        result:
          $ref: '#/components/schemas/TopicAnalysisResult'
    TopicAnalysisProgress:
      type: object
      properties:
        startedAt:
          type: integer
          format: int64
        completenessPercent:
          type: number
          format: decimal
        msgsScanned:
          type: integer
          format: int64
        bytesScanned:
          type: integer
          format: int64
    TopicAnalysisResult:
      type: object
      properties:
        startedAt:
          type: integer
          format: int64
        finishedAt:
          type: integer
          format: int64
        error:
          type: string
        totalStats:
          $ref: '#/components/schemas/TopicAnalysisStats'
        partitionStats:
          type: array
          items:
            $ref: '#/components/schemas/TopicAnalysisStats'
    TopicAnalysisSizeStats:
      type: object
      properties:
        sum:
          type: integer
          format: int64
        min:
          type: integer
          format: int64
        max:
          type: integer
          format: int64
        avg:
          type: integer
          format: int64
        prctl50:
          type: integer
          format: int64
        prctl75:
          type: integer
          format: int64
        prctl95:
          type: integer
          format: int64
        prctl99:
          type: integer
          format: int64
        prctl999:
          type: integer
          format: int64
    TopicAnalysisStats:
      type: object
      properties:
        partition:
          type: integer
          format: int32
          description: null if this is total stats
        totalMsgs:
          type: integer
          format: int64
        minOffset:
          type: integer
          format: int64
        maxOffset:
          type: integer
          format: int64
        minTimestamp:
          type: integer
          format: int64
        maxTimestamp:
          type: integer
          format: int64
        nullKeys:
          type: integer
          format: int64
        nullValues:
          type: integer
          format: int64
        approxUniqKeys:
          type: integer
          format: int64
        approxUniqValues:
          type: integer
          format: int64
        keySize:
          $ref: '#/components/schemas/TopicAnalysisSizeStats'
        valueSize:
          $ref: '#/components/schemas/TopicAnalysisSizeStats'
        hourlyMsgCounts:
          type: array
          items:
            type: object
            properties:
              hourStart:
                type: integer
                format: int64
              count:
                type: integer
                format: int64
...
servers:
  - url: http://localhost:8080
    description: Default endpoint for Kafbat UI API
    variables: {}